+++ run.py
@@ -80,16 +80,16 @@ def install_deps(
     build_dir: str,
     install_dir: str,
     debug: bool,
-    webrtc_build_dir: Optional[str],
-    webrtc_build_args: List[str],
-    sora_dir: Optional[str],
-    sora_args: List[str],
+    local_webrtc_build_dir: Optional[str],
+    local_webrtc_build_args: List[str],
+    local_sora_cpp_sdk_dir: Optional[str],
+    local_sora_cpp_sdk_args: List[str],
 ):
     with cd(BASE_DIR):
         version = read_version_file("VERSION")
 
         # WebRTC
-        if webrtc_build_dir is None:
+        if local_webrtc_build_dir is None:
             install_webrtc_args = {
                 "version": version["WEBRTC_BUILD_VERSION"],
                 "version_file": os.path.join(install_dir, "webrtc.version"),
@@ -102,16 +102,19 @@ def install_deps(
         else:
             build_webrtc_args = {
                 "platform": platform,
-                "webrtc_build_dir": webrtc_build_dir,
-                "webrtc_build_args": webrtc_build_args,
+                "local_webrtc_build_dir": local_webrtc_build_dir,
+                "local_webrtc_build_args": local_webrtc_build_args,
                 "debug": debug,
             }
 
             build_webrtc(**build_webrtc_args)
 
-        webrtc_info = get_webrtc_info(platform, webrtc_build_dir, install_dir, debug)
+        webrtc_info = get_webrtc_info(platform, local_webrtc_build_dir, install_dir, debug)
 
-        if platform in ("ubuntu-20.04_x86_64", "ubuntu-22.04_x86_64") and webrtc_build_dir is None:
+        if (
+            platform in ("ubuntu-20.04_x86_64", "ubuntu-22.04_x86_64")
+            and local_webrtc_build_dir is None
+        ):
             webrtc_version = read_version_file(webrtc_info.version_file)
 
             # LLVM
@@ -167,10 +170,16 @@ def install_deps(
             add_path(os.path.join(install_dir, "cmake", "bin"))
 
         # Sora C++ SDK
-        if sora_dir is None:
+        if local_sora_cpp_sdk_dir is None:
             install_sora_and_deps(platform, source_dir, install_dir)
         else:
-            build_sora(platform, sora_dir, sora_args, debug, webrtc_build_dir)
+            build_sora(
+                platform,
+                local_sora_cpp_sdk_dir,
+                local_sora_cpp_sdk_args,
+                debug,
+                local_webrtc_build_dir,
+            )
 
         # CLI11
         install_cli11_args = {
@@ -226,10 +235,10 @@ def main():
     )
     parser.add_argument("--debug", action="store_true")
     parser.add_argument("--relwithdebinfo", action="store_true")
-    parser.add_argument("--webrtc-build-dir", type=os.path.abspath)
-    parser.add_argument("--webrtc-build-args", default="", type=shlex.split)
-    parser.add_argument("--sora-dir", type=os.path.abspath)
-    parser.add_argument("--sora-args", default="", type=shlex.split)
+    parser.add_argument("--local-webrtc-build-dir", type=os.path.abspath)
+    parser.add_argument("--local-webrtc-build-args", default="", type=shlex.split)
+    parser.add_argument("--local-sora-cpp-sdk-dir", type=os.path.abspath)
+    parser.add_argument("--local-sora-cpp-sdk-args", default="", type=shlex.split)
     parser.add_argument("--package", action="store_true")
 
     args = parser.parse_args()
@@ -250,10 +259,10 @@ def main():
         build_dir,
         install_dir,
         args.debug,
-        args.webrtc_build_dir,
-        args.webrtc_build_args,
-        args.sora_dir,
-        args.sora_args,
+        args.local_webrtc_build_dir,
+        args.local_webrtc_build_args,
+        args.local_sora_cpp_sdk_dir,
+        args.local_sora_cpp_sdk_args,
     )
 
     configuration = "Release"
@@ -264,9 +273,11 @@ def main():
 
     mkdir_p(os.path.join(build_dir, "zakuro"))
     with cd(os.path.join(build_dir, "zakuro")):
-        webrtc_info = get_webrtc_info(platform, args.webrtc_build_dir, install_dir, args.debug)
+        webrtc_info = get_webrtc_info(
+            platform, args.local_webrtc_build_dir, install_dir, args.debug
+        )
         webrtc_version = read_version_file(webrtc_info.version_file)
-        sora_info = get_sora_info(platform, args.sora_dir, install_dir, args.debug)
+        sora_info = get_sora_info(platform, args.local_sora_cpp_sdk_dir, install_dir, args.debug)
 
         with cd(BASE_DIR):
             version = read_version_file("VERSION")